Document GNU_PROPERTY_X86_XXX
The .note.gnu.property section with NT_GNU_PROPERTY_TYPE_0 has been
added to Linux Extensions to gABI:

https://gitlab.com/x86-psABIs/Linux-ABI.git

GNU_PROPERTY_X86_XXX are processor-specific program property types for
i386 and x86-64.

1. GNU_PROPERTY_X86_UINT32_AND_LO..GNU_PROPERTY_X86_UINT32_AND_HI
A bit in the output pr_data field is set only if it is set in all
relocatable input pr_data fields.  If all bits in the the output
pr_data field are zero, this property should be removed from output.

If the bit is 1, all input relocatables have the feature.  If the
bit is 0 or the property is missing, the info is unknown.

2. GNU_PROPERTY_X86_UINT32_OR_LO..GNU_PROPERTY_X86_UINT32_OR_HI
A bit in the output pr_data field is set if it is set in any
relocatable input pr_data fields. If all bits in the the output
pr_data field are zero, this property should be removed from output.

If the bit is 1, some input relocatables have the feature.  If the
bit is 0 or the property is missing, the info is unknown.

3.  GNU_PROPERTY_X86_UINT32_OR_AND_LO..GNU_PROPERTY_X86_UINT32_OR_AND_HI
A bit in the output pr_data field is set if it is set in any relocatable
input pr_data fields and this property is present in all relocatable
input files.  A missing property implies that its bits have unknown
values.  When all bits in the output pr_data field are zero, this
property should not be removed from output to indicate it has zero in
all bits.

If the property is in output, all input relocatables have the property.
If the bit is 1, some input relocatables have the feature.  If the
bit is 0, none of input relocatables have the feature.

GNU_PROPERTY_X86_ISA_1_NEEDED can be used to describe micro-architecture
levels contained in x86 ELF binaries.