APX: Add R_X86_64_CODE_4_GOTTPOFF
For TLS instructions:

	mov	foo@gottpoff(%rip), %reg
	add	foo@gottpoff(%rip), %reg

add

 #define R_X86_64_CODE_4_GOTTPOFF	44

if the instruction starts at 4 bytes before the relocation offset.  This
should be used if reg is one of the additional general-purpose registers,
r16-r31, in Intel APX.  It is similar to R_X86_64_GOTTPOFF and linker
optimization must take the different instruction encoding into account.