va_arg: max fp_offset is 176
The document still used 304.  When we designed the psABI,
we started with all XMM registers to be usable to pass arguments
(and hence the max fp_offset was 6*8+16*16).  In 3c908ac0 (August
2001) we switched to the final scheme, using only 8 XMM registers,
but failed to adjust the offsets in the va_arg machinery.

This change is editorial only, it can't affect any real-world
compiler, as using 304 would have been incompatible with any
existing one.